requirements.txt
absl-py
numpy>=1.8.0rc1
six>=1.4.1
